+++ examples/library_usage_examples.py
@@ -1,13 +1,14 @@
 #!/usr/bin/env python
 # coding: utf-8
 
+# BEGIN --- required only for testing, remove in real world code --- BEGIN
 import os
 import sys
-
 THISDIR = os.path.dirname(os.path.abspath(__file__))
-# Add app dir to path for testing pyjsonrpc
 APPDDIR = os.path.abspath(os.path.join(THISDIR, os.path.pardir))
 sys.path.insert(0, APPDDIR)
+# END --- required only for testing, remove in real world code --- END
+
 
 import pyjsonrpc