+++ history.rst
@@ -6,6 +6,16 @@ by Gerold Penz 2013
 
 
 =============
+Version 0.2.5
+=============
+
+2013-06-30
+
+- Now, it is possible to use the *attribute* name as *method* name for
+  HTTP-JSON-RPC Requests.
+
+
+=============
 Version 0.2.4
 =============